requirements.txt
pandas==1.5.3
numpy
scikit-learn
beautifulsoup4==4.12.0
html5lib
ipdb==0.13.13
html-table-parser-python3==0.3.1
Omegaconf
plotly
plotly-express
kaleido
sphinx
sphinx_rtd_theme
hydra-core
einops